gpt_project_v3/requirements.txt
streamlit==1.29.0
PyPDF2==1.26.0
langchain==0.0.350  
transformers==4.12.0
torch==1.10.0
python-dotenv==1.0.0